+++ DeepSpaceSaga.Common/Universe/Entities/CelestialObjects/Spacecrafts/ISpacecraft.cs
@@ -5,7 +5,7 @@ public interface ISpacecraft : ICelestialObject
     float MaxSpeed { get; set; }
     float Agility { get; set; }
     List<IModule> Modules { get; set; }
-    IModule GetModule(int moduleId);
+    IModule GetModule(long moduleId);
     List<IModule> GetModules(Category category); 
     void SetDirection(double direction);
 }
